Force type conversion to avoid compilation errors on Arduino Leonardo and Micro when writing 0x00 using Serial.write(0x00).

+++ REXduino_slave/REXduino_commands.ino
@@ -45,10 +45,10 @@ void commandV(byte verbose) {
   else
   {
     Serial.print("V");
-    Serial.write(MAJORVERSION);
-    Serial.write(MINORVERSION);
-    Serial.write(REVISION);
-    Serial.write(COMMIT);
+    Serial.write((byte)MAJORVERSION);
+    Serial.write((byte)MINORVERSION);
+    Serial.write((byte)REVISION);
+    Serial.write((byte)COMMIT);
     Serial.print(";");
   }
 }
